requirements.txt
pandas
openpyxl
playwright
parsel
cairosvg
pillow
twocaptcha